backend/pyproject.toml
[project]
name = "echolog"
version = "0.1.0"
description = "AI-powered customer feedback triage tool"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0,<4.0",
    "flask-sqlalchemy>=3.1,<4.0",
    "flask-migrate>=4.0,<5.0",
    "sqlalchemy>=2.0,<3.0",
    "pydantic>=2.0,<3.0",
    "structlog>=24.0",
    "openai>=1.0,<2.0",
    "anthropic>=0.30,<1.0",
    "python-dotenv>=1.0,<2.0",
    "gunicorn>=22.0,<23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-flask>=1.3,<2.0",
    "pytest-cov>=5.0,<6.0",
    "factory-boy>=3.3,<4.0",
    "ruff>=0.5,<1.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
